Add escape @@{...} functionality
I ran into an issue where the text "...@{...}..." inside of prose or
code blocks would get treated as a reference to a code block when I
didn't want it to.

In my case, I was just trying to define a regular expression.

Also adds support custom PREFIX for install so that you can
install this in a user local directory.