Add ability to link to chapter/section
To link to a chapter/section:
[anchor text](@chapter or section title)

This change adds a slot to the weaver to store a map
of chapter/section titles to their IDs.

Then when we're weaving, we search for these links and
replace them with their [anchor text](#c:id).

This commit is a work-in-progress that I want to at least
get out to start bouncing ideas around.